# optimize lu parser to improve parsing performance and robustness