--- a/packages/lu/src/parser/lufile/LUFileParser.g4
+++ b/packages/lu/src/parser/lufile/LUFileParser.g4
@@ -10,6 +10,8 @@ paragraph
     : newline
     | nestedIntentSection
     | simpleIntentSection
+    | entitySection
+    | newEntitySection
     | importSection
     | qnaSection
     | modelInfoSection
@@ -51,7 +53,7 @@ subIntentDefinition
     ;
 
 simpleIntentSection
-    : (intentDefinition? (entitySection | newEntitySection)+) | intentDefinition
+    : intentDefinition
     ;
 
 intentDefinition