--- a/packages/lu/src/parser/lufile/generated/LUFileLexer.js
+++ b/packages/lu/src/parser/lufile/generated/LUFileLexer.js
@@ -1,4 +1,4 @@
-// Generated from ../LUFileLexer.g4 by ANTLR 4.7.2
+// Generated from ../LUFileLexer.g4 by ANTLR 4.8
 // jshint ignore: start
 var antlr4 = require('antlr4/index');
 
@@ -553,7 +553,6 @@ LUFileLexer.prototype.ruleNames = [ "A", "B", "C", "D", "E", "F", "G", "H",
                                     "QNA_TEXT" ];
 
 LUFileLexer.prototype.grammarFileName = "LUFileLexer.g4";
-
 
 
 exports.LUFileLexer = LUFileLexer;