--- a/packages/lu/src/parser/lufile/generated/LUFileParserVisitor.js
+++ b/packages/lu/src/parser/lufile/generated/LUFileParserVisitor.js
@@ -1,4 +1,4 @@
-// Generated from ../LUFileParser.g4 by ANTLR 4.7.2
+// Generated from ../LUFileParser.g4 by ANTLR 4.8
 // jshint ignore: start
 var antlr4 = require('antlr4/index');