--- a/packages/lu/test/parser/lufile/sectionapi.test.js
+++ b/packages/lu/test/parser/lufile/sectionapi.test.js
@@ -30,7 +30,6 @@ describe('Section CRUD tests for intent', () => {
         assert.equal(luresource.Sections[0].UtteranceAndEntitiesMap[1].utterance, 'hello');
     });
 
-
     it('add simpleIntentSection test', () => {
         let newFileConent = 
         `# CheckEmail
@@ -194,6 +193,23 @@ describe('Section CRUD tests for intent', () => {
         assert.equal(luresource.Sections[0].Body, `> comment1${NEWLINE}- hi${NEWLINE}hello${NEWLINE}$${NEWLINE}@${NEWLINE}> comment2`);
     });
 
+    it('update section test for invalid content2', () => {
+        // missing entity type after '$a', here 'a' is entity name
+        let newFileConent = `# Greeting${NEWLINE}> comment1${NEWLINE}- hi${NEWLINE}$a${NEWLINE}> comment2`;
+
+        let sectionId = luresource.Sections[0].Id;
+        luresource = new SectionOperator(luresource).updateSection(luresource.Sections[0].Id, newFileConent);
+        assert.equal(luresource.Errors.length, 1);
+        assert.equal(luresource.Sections.length, 4);
+        assert.equal(luresource.Sections[0].Errors.length, 1);
+        assert.equal(luresource.Sections[0].SectionType, LUSectionTypes.SIMPLEINTENTSECTION);
+        assert.equal(luresource.Sections[0].Name, 'Greeting');
+        assert.equal(luresource.Sections[0].Id, sectionId);
+        assert.equal(luresource.Sections[0].UtteranceAndEntitiesMap.length, 1);
+        assert.equal(luresource.Sections[0].UtteranceAndEntitiesMap[0].utterance, 'hi');
+        assert.equal(luresource.Sections[0].Body, `> comment1${NEWLINE}- hi${NEWLINE}$a${NEWLINE}> comment2`);
+    });
+
     it('add section test for nested section content with comments', () => {
         let simpleIntentBody1 = `- hello${NEWLINE}- hi${NEWLINE}> this is comment 1${NEWLINE}@${NEWLINE}> this is comment 2${NEWLINE}`
         let simpleIntentBody2 = `- bye${NEWLINE}$${NEWLINE}> this is comment 3${NEWLINE}${NEWLINE}> this is comment 4${NEWLINE}${NEWLINE}`
@@ -202,7 +218,7 @@ describe('Section CRUD tests for intent', () => {
 
         luresource = new SectionOperator(luresource).addSection(newFileConent);
 
-        assert.equal(luresource.Errors.length, 7);
+        assert.equal(luresource.Errors.length, 4);
         assert.equal(luresource.Sections.length, 6);
         assert.equal(luresource.Sections[4].Errors.length, 3);
         assert.equal(luresource.Sections[4].SectionType, LUSectionTypes.NESTEDINTENTSECTION);