--- a/packages/lu/test/parser/lufile/sectionapi.test.js
+++ b/packages/lu/test/parser/lufile/sectionapi.test.js
@@ -420,3 +420,81 @@ hello
         assert.equal(luresource.Sections[0].Body.replace(/\r\n/g,"\n"), insertFileContent);
     });
 });
+
+describe('Section range tests', () => {    
+    it('simple intent section test', () => {
+        let fileContent =
+`# Cancel
+- cancel it
+
+#    Greeting
+## HelloGreeting
+- hi
+
+@simple e1
+
+## ByeGreeting
+- bye
+
+@ ml e2`;
+
+        let luresource = luparser.parse(fileContent);
+
+        assert.equal(luresource.Errors.length, 1);
+        assert.equal(luresource.Sections.length, 4);
+
+        assert.equal(luresource.Sections[0].SectionType, LUSectionTypes.SIMPLEINTENTSECTION);
+        assert.equal(luresource.Sections[0].Name, 'Cancel');
+        assert.equal(luresource.Sections[0].Range.Start.Line, 1)
+        assert.equal(luresource.Sections[0].Range.Start.Character, 0)
+        assert.equal(luresource.Sections[0].Range.End.Line, 3)
+        assert.equal(luresource.Sections[0].Range.End.Character, 0)
+
+        assert.equal(luresource.Sections[1].SectionType, LUSectionTypes.SIMPLEINTENTSECTION);
+        assert.equal(luresource.Sections[1].Name, 'Greeting');
+        assert.equal(luresource.Sections[1].Range.Start.Line, 4)
+        assert.equal(luresource.Sections[1].Range.Start.Character, 0)
+        assert.equal(luresource.Sections[1].Range.End.Line, 4)
+        assert.equal(luresource.Sections[1].Range.End.Character, 13)
+
+        assert.equal(luresource.Sections[2].SectionType, LUSectionTypes.SIMPLEINTENTSECTION);
+        assert.equal(luresource.Sections[2].Name, 'HelloGreeting');
+        assert.equal(luresource.Sections[2].Range.Start.Line, 5)
+        assert.equal(luresource.Sections[2].Range.Start.Character, 0)
+        assert.equal(luresource.Sections[2].Range.End.Line, 9)
+        assert.equal(luresource.Sections[2].Range.End.Character, 0)
+
+        assert.equal(luresource.Sections[3].SectionType, LUSectionTypes.SIMPLEINTENTSECTION);
+        assert.equal(luresource.Sections[3].Name, 'ByeGreeting');
+        assert.equal(luresource.Sections[3].Range.Start.Line, 10)
+        assert.equal(luresource.Sections[3].Range.Start.Character, 0)
+        assert.equal(luresource.Sections[3].Range.End.Line, 13)
+        assert.equal(luresource.Sections[3].Range.End.Character, 7)
+    });
+
+    it('nested intent section test', () => {
+        let fileContent =
+`> !# @enableSections = true
+# Greeting
+## HelloGreeting
+- hi
+
+@simple e1
+
+## ByeGreeting
+- bye
+
+@ ml e2`;
+
+        let luresource = luparser.parse(fileContent);
+        assert.equal(luresource.Errors.length, 0);
+        assert.equal(luresource.Sections.length, 2);
+
+        assert.equal(luresource.Sections[1].SectionType, LUSectionTypes.NESTEDINTENTSECTION);
+        assert.equal(luresource.Sections[1].Name, 'Greeting');
+        assert.equal(luresource.Sections[1].Range.Start.Line, 2)
+        assert.equal(luresource.Sections[1].Range.Start.Character, 0)
+        assert.equal(luresource.Sections[1].Range.End.Line, 11)
+        assert.equal(luresource.Sections[1].Range.End.Character, 7)
+    });
+})