This PR is fixing two issues:

1. Follow up PR of #905. In this PR, we adjust grammar file of LU parser to improve the parsing performance.
Now the parsing performace will be stable for lu files with all kinds of formats. The performance table is listed as below:

    File Size | Before | After | After(with cache)
   -- | -- | -- | --
   1100 lines | 700ms | 150ms | 120ms
   250 lines | 350ms | 120ms | 80ms

2. Fix #907. Before the fix, entity line of '$a' after a intent will not be treated as entity definition. After the fix, it will be recognized as entity section with an error like 'no entity type'. This will help composer lu editor to parse and keep the invalid entity definition.